I am the treelord. also basic materials on the level, gonna brush over it when it isnt 2am lmao
i dont even know why people read these descriptions, if they do that is. they are honestly the most inaccesible things to me. like come on, you would think an explanation of what the last push is should be as visible as possible but nope lets just it at hidden in some weird font inside the header that can just as easily be made the description of what you did. Wow! speaking of which, look at what i did! Amazing! how commendable of me to cut out the tedious middleman provided only by github. thanks assholes. The only reason im typing this is because im almost positive no one will read this and im in too much of a fugue state to stop myself. but yeah look at the title to see what i did, and also discord because communication is a thing.